Update bit_decompose def:
- Move bit_decompose into hypercube.rs which is closer conceptually and
  only used there (except for a VirtualPolynomial test)
- Remove unused bit_decompose imports from other files
