+++ src/utils.rs
@@ -196,17 +196,6 @@ pub fn hadamard_product<F: PrimeField>(a: &Vec<F>, b: &Vec<F>) -> Vec<F> {
   res
 }
 
-/// Decompose an integer into a binary vector in little endian.
-pub fn bit_decompose(input: u64, num_var: usize) -> Vec<bool> {
-  let mut res = Vec::with_capacity(num_var);
-  let mut i = input;
-  for _ in 0..num_var {
-    res.push(i & 1 == 1);
-    i >>= 1;
-  }
-  res
-}
-
 /// Sample a random list of multilinear polynomials.
 /// Returns
 /// - the list of polynomials,
